+++ internals/download-manual.js
@@ -1,10 +1,10 @@
 const fs = require('fs')
-const got = require('got')
+const fetch = require('node-fetch')
 
-got
-  .get(
-    'https://raw.githubusercontent.com/stedolan/jq/master/docs/content/manual/manual.yml',
-  )
-  .then(({ body }) => {
-    fs.writeFileSync('./files/manual.yml', body)
+fetch(
+  'https://raw.githubusercontent.com/stedolan/jq/master/docs/content/manual/manual.yml',
+)
+  .then((res) => res.text())
+  .then((text) => {
+    fs.writeFileSync('./files/manual.yml', text)
   })